requirements.txt
Flask==3.0.3
groq==0.5.0
mediapipe
numpy==1.26.4
opencv-python==4.9.0.80
openai==1.24.1
requests==2.31.0
fastapi==0.110.1
uvicorn[standard]==0.29.0
pydantic==2.7.0
python-multipart==0.0.9
python-dotenv==1.0.1
